requirements.txt
pandas==1.5.1
python-dotenv==1.0.1
Requests==2.32.3
selenium==4.22.0
selenium_wire==5.1.0
